Add navigation between resumes list, button create resume and button view resume detail
