--- src/components/resumes/ResumeList/ResumeListAPI.js
@@ -1,21 +1,0 @@
-import axios from "axios";
-
-export default async function getResumes() {
-  let handlerError = false;
-  try {
-    const { data: result } = await axios.get(
-      "https://jsonplaceholder.typicode.com/users"
-    );
-    return result;
-  } catch (error) {
-    if (error.response) {
-      alert(error);
-    } else if (error.request) {
-      alert(error);
-    } else {
-      alert("Error: Something is wrong");
-    }
-    handlerError = true;
-    return handlerError;
-  }
-}
