+++ src/components/resumes/ResumesAPI.js
@@ -19,3 +19,23 @@ export async function postResume(resume) {
     return { ok: false };
   }
 }
+
+export default async function getResumes() {
+  let handlerError = false;
+  try {
+    const { data: result } = await axios.get(
+      "http://localhost:7071/api/v1/resumes"
+    );
+    return result;
+  } catch (error) {
+    if (error.response) {
+      alert(error);
+    } else if (error.request) {
+      alert(error);
+    } else {
+      alert("Error: Something is wrong");
+    }
+    handlerError = true;
+    return handlerError;
+  }
+}
